Added automatic calculation of period
Removed the user input and interaction for finding the  period. Added a function that automatically calculates the period in its place. Also added an array for checking the IC values and their likely correlating periods (currently limited to 20). Added a function that uses that table and a calculated IC period to estimate the suggested period. Added/Edited some comments for clarity.