Vigenere Code Rewrite
Rewrote the vigenere program to more easily allow future improvements.
Made every function static.
Added a function that automatically uses the new functions to completely solve a vigenere cipher.
Changed period calculation to use a different method and return a list of possible periods.
Changed decrypting to return a string instead of outputting, to do the whole text at once, and to keep extra characters, spaces, and letter capitalization.
Changed the remove spaces function to remove all special characters also.
Made a new function that returns the index of the largest value in an array.
Made a function to calculate an alphabet's likely keys and removed the old function that stored translated alphabets.
Made a function to find the most likely overall key from the calculated keys.
Made a function to turn the key into a string.
Modified the run file to work after the changes.