Finished vigenere menu functionality
Added the following functionality to the vigenere menu:
Changing to the next most likely period.
Having the user set the period.
Changing to the next most likely key for an alphabet.
Having the user set the key for an alphabet.
Having the user set the whole key for the text.
Outputting the current solution to a file.

Changed the vigenere run function to use the menu.
Added a function that gets a key from a user.
Added a function to convert a key from a string to an integer array.
Added a function to get an integer input from the user in a specified range.
Added a function to get a single letter input from the user.

Changed outputs to be more consistent throughout.