Added vigenere encryption
Added a function that takes a text and a key and encrypts it with a vigenere cipher.
Added a function that asks the user for a key and text and then encrypts with it.
Renamed VigenereSolver to Vigenere.
Changed the vigenereRun file to ask if the user wants to encrypt or decrypt.