Added caesar encryption
Added a function that takes a text and a shift and encrypts it with a caesar cipher.
Added a function that asks the user for a shift and text and then encrypts with it.
Renamed CaesarSolver to Caesar.
Changed the CaesarRun file to ask if the user wants to encrypt or decrypt.