Added solveCaesar function to Caesar
- Some of the old functions of Caesar may be broken this build

Added a solveCaesar function for Caesar that does what the constructor of Caesar does and then prints the results.
Changed getFrequencyOfLetters to not include spaces and work with lowercase letters.
Modified getCorrelationOfFrequencies to work with strings without spaces.
Modified printResults to be static.
Changed unshift to work with punctuation and lowercase letters.

Changed CaesarRun to use the new solveCaesar function.
This should run the same as the last build except it will keep all characters and capitalization.

Added deallocation of memory.
Added some comments.